Created WaterProgressTrackService

+++ watercat/Model/IWaterNotificationService.cs
@@ -4,5 +4,5 @@ public interface IWaterNotificationService
 {
     Task RequestPermissionAsync();
     void SendNotification(string title, string msg, int id = 100);
-    void SendWaterProgressNotification(double percent, int remaining);
+    void SendWaterProgressNotification(double percent, double remaining);
 }
